Refactor to let unit tests run without dependency
