+++ .travis.yml
@@ -3,6 +3,7 @@ python:
   - "2.7"
 cache: pip
 install:
+  - python setup.py install
   - pip install -r requirements-dev.txt
 script:
-  - nosetests
+    - nosetests --exclude-dir="test/disco_tests"
