+++ requirements-dev.txt
@@ -1,2 +1,3 @@
 mock==2.0.0
 nose==1.3.7
+nose-exclude==0.5.0
